Improve chash equality complexity and narrow default offset type
* Rewrite operator== from O(N^2) (triple nested scan) to O(N): the unique
  branch walks the LHS once and probes the RHS via find / find_value_with_hash_,
  while the multi branch groups by equal_range and compares as a multiset.
  Tag dispatch picks between the two without if constexpr, and an SFINAE
  check on `get_hasher() == rhs.get_hasher()` enables hash reuse only when
  both sides share hasher state, falling back to recomputing through find on
  the RHS otherwise.
* Factor the bucket walk into find_value_with_hash_, so the new equality
  path and existing find share one implementation.
* Extend the move-tag hierarchy to match bpptree / segment_array
  (move_noexcept_tag / move_assign_tag : move_base_tag) and add nothrow-move
  overloads for the relocation helpers; merge destroy_one onto the parent
  tag where the two paths were identical.
* Drop the default offset_type in chash_map.h and chash_set.h from
  std::uintptr_t to std::uint32_t. The field is a slot index, not a pointer
  width; the previous 64-bit default wasted ~12 bytes per element on LP64
  with no extra reach, and a custom config can still widen it where needed.
* Wrap memmove destinations in static_cast<void *> to clear
  -Wclass-memaccess; chash_compile.cpp / chash_test.cpp adapted.

+++ chash_map.h
@@ -11,7 +11,7 @@ struct chash_map_config_t
     typedef hasher_t hasher;
     typedef key_equal_t key_equal;
     typedef allocator_t allocator_type;
-    typedef std::uintptr_t offset_type;
+    typedef std::uint32_t offset_type;
 #if __cplusplus >= 201703L
     typedef typename std::invoke_result<hasher, key_type>::type hash_value_type;
 #else
